Ensure KotlinSpecPropDefault Assertions
Summary:
Now that the assertions are optional, it could actually be pretty easy to not run the actual assertion lambda.  Right now, nothing would fail if that occurred.

To make this more robust to potential changes, this adds an automic boolean that we can utilize to ensure that we always run the assertion.  Now, if we ended up in a case where an assertion was not run for some reason, we won't silently fail.

Reviewed By: adityasharat

Differential Revision: D34491996

fbshipit-source-id: 3af87508a5693bd6184ccb2b941c20196c4421c9