Move useTransition docs into hooks intro for spec developers
Summary: useTransition should be covered in the animations basics so I don't want to move it to mainconcepts. Since it mostly is a code example of before/after from specs, I've just moved it into the 'hooks for spec developers' page. We'll need to follow up with better animations docs.

Reviewed By: SidharthGuglani

Differential Revision: D34523197

fbshipit-source-id: 5bb0eece2eb9fcecc483822132585328b4c1442d