Removed IM enabled requirement from IncrementalMountUtils
Summary: IncrementalMountUtils only works on LithoViews that have IM enabled. However, by forcing this condition, we prevent the normal behaviour for processing visibility outputs when IM is disabled. notifyVisibleBoundsChanged already checks if IM is enabled, and if not, will simply process the visibility outputs. By removing this condition from IMU, we allow for this normal behaviour to take place.

Reviewed By: adityasharat

Differential Revision: D34642671

fbshipit-source-id: d2da1a94f82452dad1cc686ba75b6943b32eaa7e