Fix all of the differences after the widget move to kotlin package
Summary:
I definetly shouldve address all of those in the previous diff, but landing and rebasing was a lttle bit challenging :(
So i tried to land it without any failures first...
Now after slowly going through all of the changed files, i can see that during rebasing I intrduced some changes.
This diff is bringing all of the code to how it was and removes any additional targets that i added during rebase.

Reviewed By: astreet

Differential Revision: D34680742

fbshipit-source-id: 8ec773694538ab78a2142075a3ba7e024fa1caae