Apply state update early
Summary:
Extracting out the applying state before render method. We iterate through the pendingStateUpdates and apply state update on cloned StateContainer.

Also we are keeping the StateContainers for main tree in StateHandler and in Initial State Container for nested tree as we don't want to mutate StateHandler during layout time.

Initial State creation which also be moved out of render flow once we implement an API to resolve all nested tree together.

Reviewed By: adityasharat

Differential Revision: D34546923

fbshipit-source-id: 9b2bbefd79746c23783e9ba42816905d804f70d4