Add Ability to Find/Get Root Component
Summary:
This diff adds the ability to easily get the root component or head component.  This adds an easy way to get the root or head component, which is something that was missing from the new API.  (At least, an easy way to do it.)

## Implementation Notes:
If we don't want to use the scoped component infos and get the head, we could try to get the root from the component tree instead.  The benefit I see in having this function, however, in the ComponentsFinder is that you can refer back to it when needing to find direct components.

## More Largely

More largely, we want to:
1. **Add the ability to get the `headComponent`**
2. ~~Ensure when finding components we search from head to tail (and not tail to head, as we have now)~~
3. Update docs for `findDirectComponent` to explain how direct components includes a list
4. Can add `hasDirectComponent` as a helper/nice assertion to ease things via non-null checks
5. ^perhaps also `doesNotHaveDirectComponent`
6. Update finding matching direct component to match against all direct components

------------

Reviewed By: adityasharat

Differential Revision: D34599337

fbshipit-source-id: 5816a868a5d769b44d8e42aff7b01ec442185ce8