During unmount, call unbind only after the item has been unmounted from the host
Summary: When unmounting, we are currently calling unbind on extensions, and only then removing the item from the host. During unbind, we unset several properties, including dynamic properties such as background, scale, translation, etc. When setting such properties on a view that's still in the view hierarchy, it invokes things like invalidate and requestLayout all the way up the view tree. Since we're about to be removing the view from the hierarchy in the next line of code anyway, this step is redundant, and just adds extra perf overhead.

Reviewed By: adityasharat

Differential Revision: D34821988

fbshipit-source-id: 84f327ef86c7c49ff57092622366b9fb2d5d52d1