Ensures that view attribute flags for the LithoView is from initialisation
Summary:
Ensures that view attribute flags for the LithoView is from initialisation.
Since a LithoView's view attributes could be modified from outside of litho, and in order correctly unset the view attributes (like enabled, clickable, etc)  for the root LithoView during unmount the flags for the root LithoView needs to be captured when the view is created.

Reviewed By: mihaelao

Differential Revision: D34974966

fbshipit-source-id: 848d6f8888268dd0863cbe5dc97534520b8d0654