Add `setLithoViewLayoutParams ` to `RecyclerBinderViewHolder`
Summary:
This is another approach of customising LithoView's layout in RecyclerBinder.
We can assume that customers don't want to use the default layout provided by RecyclerBinder if they have chosen using their own `RecyclerBinderViewHolder`. We can't determine what LayoutParams should be applied to LithoView from the RecyclerBinder's point of view util csutomers tell us, so need this method to be implemented.
In comparison to the other option that using `LayoutInfo`, the advantages are listed below:
- Customers could customise multiple different layouts for ViewHolders, and single layout for all ViewHolders is just a special use case.
- Customers don't have to care about `LayoutManager` and `LayoutInfo`.
- Change is less.

Reviewed By: mihaelao

Differential Revision: D35469660

fbshipit-source-id: 8918245e35266b447c4d1f767197e3d4b043bc1e