Fixes the issues padding and border
Summary:
Fixes the issues padding and border which is cause by merging measure and bound defined calls for MountableComponents. This is done by subtracting the padding and border explicitly.

To note, this fix perceptually causes a difference when border is used with explicit width and height (see 3rd row in test plan) because MountSpec implementation doesn't subtract the border. This is the correct behaviour, and the trade off is worth it because without the fix it would be break cases where explicit width and height are not set, which is more common.

Differential Revision: D35928217

fbshipit-source-id: bb11346f787d176b5e21414dfb9745ed099c6566