Do not create shallow copy of StateHandler for lazy state updates
Summary:
We are applying lazy state update to newly created StateContainer - so that we can ensure that State params passed to Events is latest. This updated state container is not updated in StateHandler or ISC.
StateHandler or ISC still contains the state which was before applying lazy state update.

Therefore, We don't need to create shallow copy here as we just apply lazy state updates on copied StateHandler which can be done on original StateHandler.

Reviewed By: adityasharat

Differential Revision: D35962275

fbshipit-source-id: 019f250ded4630044bdd53bf666c89e3c10a9184