Throw soft error only if component is actually present in tree hierarchy
Summary:
If we remove a component from Component hierarchy and there is some state update on this component then we should not crash.

We are here checking if component is actually present in hierarchy for which we had state update and could not find the StateContainer while applying state update early. If component is present in hierarchy then we emit soft error.

Reviewed By: pasqualeanatriello

Differential Revision: D35965438

fbshipit-source-id: a3e642a5dfd741a55b939b92769df5988de73a6e