Fix idle() in testing api
Summary:
Prior to this change if there was a crash in background thread, the test just waited and eventually timed out on the main thread. This diff changes the logic so that in case of an exception in background thread, the exception will be immediately propagated to the main thread.
Also the `LithoViewRuleTest` was converted from Java to Kotlin in order to be able to use `KComponents` API to write a test for this change.

Reviewed By: apowolny

Differential Revision: D36072152

fbshipit-source-id: 63376572020c1478dc90e08f5a5abdbb88d9d214