Add cache for LithoNode for LayoutWithSizeSpec in LayoutState
Summary:
This diff just adds a cache for LithoNodes resolved for LayoutWithSizeSpec. We map globalKey + widthSpec + heightSpec to LithoNode where width / height are the specs with which LayoutWithSizeSpec was resolved.

Width / height spec in the key allows us to reuse LithoNode only if LithoNode was calculated with same width / height spec. If specs have changed then we need calculate LithoNode again since we don't know how OnCreateLayoutWithSizeSpec is implemented and what will it return if specs have changed.

Reviewed By: adityasharat

Differential Revision: D36161994

fbshipit-source-id: 831ff939ca0cafa5e651ea7366ea23fb6b262861