Ensure ComponentHost invalidates if attaching view during layout
Summary:
`addViewInLayout` is used to mount Views without requesting layout on the parent host. This appears to leave the RenderNode of the host with a cached drawable state where it doesn't correctly draw its content in some cases.
This tests the effect of manually invalidating the view to flush the drawing cache.

Reviewed By: astreet

Differential Revision: D36371204

fbshipit-source-id: 58884fcb1a4f804dc8eb9c165d6d2478d496503b