Enforcing scroll target is always reached.
Summary:
## Problem

This task aims to rethink the scroll request system to remove an error situation where we can perform a scroll to a given position after the item list structure has changed. Thus, we will scroll to the wrong item.

The described situation is as follows:

https://www.internalfb.com/intern/px/p/24ths

## Proposed Solution

In the proposed solution, we won't change the flow of interactions: the scroll request will still be started at the `SectionTree` and then follow down to the `RecyclerBinder`. The big difference is that we will keep track of the pending scroll/focus request until we can execute it. This way, we can update the target position appropriately if there is any update between the moment the scroll is requested and executed.

Below you find an overall diagram of what is happening:

https://www.internalfb.com/intern/px/p/24Jzw

Reviewed By: mihaelao

Differential Revision: D36482263

fbshipit-source-id: 4d5175ba6dd547527c430146f04793dacb6ad5e0