Fix the ANR due to a deadlock
Summary: We added `addOp` in `RecyclerBinder` to identify the root cause of `IndexOutOfBoundsException`, but it would cause a deadlock in some cases. Since we may have located where the problem is, temporarily fix it by avoiding sync the whole method and moving it behind the gate, will be addressing this issue in a following diff.

Reviewed By: astreet

Differential Revision: D36931969

fbshipit-source-id: 5d1bcd52e28d2b956bdde8ce171251578502cc0f