Report focus requests happening outside main thread
Summary: We would be able to improve our scrolling requests if all are being done on the UI thread. Given the amount of client usages we have, it is hard to understand if any current usage is being done on non-UI threads. Therefore we will be logging any focus requests started on non UI-Threads, so that we can learn about the current behaviors.

Reviewed By: mihaelao

Differential Revision: D36805545

fbshipit-source-id: 0a5a8175fabf2b56c6a9103a20a0d97d23baafa4