Remove generic arg from pools, and nothing of value was lost
Summary:
This is blocking Kotlin migrations.
The current Litho processor forces `OnCreateMountContentPool` methods to use a `MountContentPool` with no generics args. This cannot work for Kotlin.

Turns out, we don't do much with these generics. There's on spot that actually kind of uses them, the rest just ... mmm.. ignore them.

So I'm opting for the easy solutions. Generic that don't work naturally should not be there. We don't need them to save one implicit cast in the code.

Reviewed By: apowolny

Differential Revision: D36984933

fbshipit-source-id: 81196a66bea2949ed059268f9905e64ee9a05a16