Add accessibility support to RCTextView ClickableSpans
Summary:
In order to support accessibility in clickable spans, `RCTextView` creates a tree of virtual accessibility nodes for clickable spans and assigns `"Button"` role to each of them by default.

This change extends the accessibility capabilities of `RCTextView` ClickableSpans by introducing a `RCAccessibleClickableSpan` type that. If `RCTextView` detects that it deals with such accessible clickable span type, it does an additional setup on the accessibility node where it sets proper contentDescription and role.

Reviewed By: rooju, nicous

Differential Revision: D36894997

fbshipit-source-id: 4e565d91e803d8932a43b3a524c6849e99ad000e