actively checking activity state in side bar component to decide whether to trigger bottom sheet
Summary:
https://fb.workplace.com/groups/litho.support/permalink/3261977487456930/

This diff is the short term solution to address above issue.

Discussed with astreet, xiaonoxus and until the lifeCycleProvider in News Feed is ready we can use this short term solution to achieve that when feed unit is on background we do not trigger the bottom sheet showing up twice.

In this diff:

And inside the subscription, actively using the valid activity to check the lifecycle current state. When it's on resume(focused/visible) then trigger the bottom sheet.

Reviewed By: astreet

Differential Revision: D37200943

fbshipit-source-id: c2ddae3b12d928aef3e2a64bd3ec0a8a8a9d6968