Don't wait on interrupted LayoutStateFutures from sync, background layouts
Summary:
The attached crash can happen if a sync bg layout uses a LayoutStateFuture which the main thread interrupts and transfers to its own thread - this causes the LSF to return null on the bg thread. This diff should add more complete protection against these cases:

- you start an async layout, then a sync layout in the background blocks on it, and then a third layout on the main thread blocks on it causing it to interrupt and transfer.
- you could also order the other way: async layout, main thread waits on it and interrupts, before that completes bg thread tries to wait on it too

The approach is to not allow a LSF to be interrupted if a sync bg layout is waiting on it, and to not let a sync bg layout wait on an interrupted layout.

Reviewed By: mihaelao

Differential Revision: D37068529

fbshipit-source-id: b51f07003bbf1b1beb838a58652170ae5cd87c4e