Schedule Async State Updates on Main Thread
Summary:
# Problem
In the current implementation we don't have a deterministic approach to state batching. This may incur in unnecessary layout calculations, when we would be better off in gathering all state updates and then commit to one single layout calculation.

# Approach
This first attempt is based on the hypothesis that we might avoid not needed layout calculations by attempting to start the calculation as soon as the *Main Thread* finishes processing its current stream of work. This way, we could batch more state updates.

This is achieved by posting a `Runnable` at the front of the *Main Thread* queue as we receive the state update. This `Runnable` will simply call the `updateStateInternal`, which will then start the layout calculation process.

If we have a series of state updates, like show below, we will be:
1. removing any pending scheduling (that should be in front of the queue)
2. adding a new "state update scheduling" in the front of the queue.

This is done because we have no way to understand if the first "state update scheduling" executed or not. So, we act defensively by always re-posting to the front of the queue.

```
onClick = {
   state1.update(1)
   state2.update(2)
   state3.update(3)
}
```

## Schema

In the current implementation, you can see that every state update schedules a layout calculation, which may or may not (depending on the schedulers) trigger before all state updates are considered. This can be bring the then not necessary additional work.

{F740499028}

In the new approach, you can see that every state update creates a "state update" scheduling in the *Main Thread* (pushing "Message 2" to the back). This decreases a lot the work on the background thread. However, there is a risk that we are creating too much delay.

{F740499301}

## Considerations
- At this moment we are only taking into account async hook updates, but adding non-hook updates would be trivial.
- This is just the base to experiment on. I'm working on top of this to try to understand/measure possible impacts.

Reviewed By: astreet

Differential Revision: D37031747

fbshipit-source-id: db00d286bdba0fe7ea20050cf35b6a3fc4c0d284