Replace usages of AndroidJunit4 runner by LithoTestRunner
Summary:
In Litho's framework context the preferred test runner to use is the `LithoTestRunner`. It allows for example to specify extra `TestConfigurations` and run all the tests in those given configurations: this can be useful when we are testing the effect of different configurations across our codebase.

The main replacements were under the following modules:
- `litho-core-kotlin`
- `litho-widget-kotlin`

Reviewed By: adityasharat

Differential Revision: D37415615

fbshipit-source-id: a36dd08a049ba892031526cf9e6d91b00689610f