Add ability to throw exception when a focus request is done not in the UI Thread
Summary:
Our initial experiment to log focus requests on the UI Thread has been up for some weeks already and we have fixed the only identified wrong usage. As a next step we plan to roll-out an experiment that actually throws an exception if a focus request happening outside the Main Thread happens.

If this roll-out is succesful we can then proceed to clean-up our code related to focus requests (and all extra-machinery actually not needed when the assumption that is run always in the UI Thread is met).

Reviewed By: mihaelao

Differential Revision: D37485660

fbshipit-source-id: d4a02a42b2d60068d506fc3e4e629c19fbb48e46