Update testing libraries so that we can remove exclude_litho_manifest
Summary:
In D37415615 (https://github.com/facebook/litho/commit/c11a5bd3119d334caa005b2eed817ff9ce2cbc6c) we had to add a gradle property that allowed modules to opt out from explicitly setting the AndroidManifest in the `LithoTestRunner`. This had to be done so that all modules could use `LithoTestRunner`.

In this diff, we aim to remove that gradle property and have one single way to run the tests.

In that last diff, we had a blocker related with. It seems this is [related](https://stackoverflow.com/a/71605255) with the specific `androidx.test.core` version we had. By bumping it the problem is solved.

Reviewed By: adityasharat

Differential Revision: D37490994

fbshipit-source-id: 57aa3535473968c9def60635949fe916510a90c1