Observe Style Callbacks to trigger Layout Calculation
Summary:
The broad idea of this approach is that we add the capability of knowing when a specific callback/event has ended. For example, if we know that a `onClick/onVisibilityChanged` callback has ended, then we could potentially check if there are any uncommitted state updates and start the Layout calculation.

This would allow us to handle each sequence of the state updates in the following way:

* Each State Update would be enqueued, but no layout calculation would be scheduled.
* As the event/callback of interest end is registered, we verify if there are any uncommited state updates and no layout calculation scheduled. If those conditions are met, we would then schedule a layout calculation.

We are able to take approach that would register a special `EventHandler` that could observe events generated via `Style` lambdas by wrapping the original event handler. This would allow us to process the event as usual, and to verify if a new layout calculation was needed. Below, you can find the high-level structure of the components.

{F743958407}

You can also follow the diagram below, which exemplifies the flow of a click event and how it could potential trigger a layout calculation (when there are uncommitted updates).

{F743990612}

Reviewed By: astreet

Differential Revision: D37238458

fbshipit-source-id: a956f5426ffe28b4665eb25a9e1acf49a126fc5e