Schedule layout calculation on Frame Callback
Summary:
In this approach we are attempting to start the layout calculation using the `Choreographer` frame callbacks system.

Whenever the `Choreographer` receives a VSYNC signal, it starts a cycle to prepare the next Frame. In this cycle is goes through 3 main phases: input handling, animation and traversals (which layouts, measures and draws).

Fortunately, the `Choreographer` API provides a way to execute code during the next *animation* phase of the processing. With this knowledge, this new variant to batch state updates will schedule the layout calculation start on the Choregrapher's animation phase. This way we can guarantee that all states generated by input handling are properly enqueued before we start the layout calculation.

You can find a high-level schema of the flow below:

{F745874666}

Reviewed By: astreet

Differential Revision: D37312147

fbshipit-source-id: 9b0f63185e846d45ed75544cacc04404959597b4