Add strategy that will flush on Choreographer callback or in front of main thread
Summary: This diff focuses in adding a new batched update strategy that will schedule the state updates and take whichever one gets ready first to execute first. The two approaches to schedule the state update are by "starting in the front of the main thread queue" and "starting in the next choreographer callback".

Reviewed By: astreet

Differential Revision: D37609093

fbshipit-source-id: 4ddc24a7bcdb35ab5d82d015980a9ff7316e7a8a