Ensure IM adheres to translation, so that off-screen items are not mounted.
Summary:
When an item is off-screen due to translation, we still must make sure that visibility events aren't fired for it. With this in mind, it might not make sense to even mount an item with IM in this case. With this in mind, the "corrected" visible rect will now also include translation of all parents.

Fix also includes ensuring child litho-views get notified when a manual position change (i.e., via translation) occurs.

Reviewed By: adityasharat

Differential Revision: D37597556

fbshipit-source-id: 3974e18fa5375760de88e7865a50400b69d4915d