Throw whenever scroll is request outside the main thread in debug mode
Summary:
So that we are able to fix the current usages of this API outside of the main thread, we are turning on the ability to
throw an exception whenever a scroll request is done outside the UI Thread AND is in debug mode.

The goal is that developers can detect these situations earlier and fix them. This way the process of fixing all wrong usages
will be faster and eventually we can fully rollout.

Reviewed By: mihaelao

Differential Revision: D37750492

fbshipit-source-id: adc91ae6e319a2f8d427c6409fc0873720cfe743