Fix issue with OnUpdateStateWithTransition
Summary:
This diff fixes a issue If OnUpdateStateWithTransition is returning a Transition with key type `TransitionKeyType.LOCAL` then it does not work.

We were not setting the owner key to the Transition which was causing this issue since we use owner key to access the Transition Id in Transition Manager.

Reviewed By: astreet

Differential Revision: D37818946

fbshipit-source-id: 5c061d8a6d3adcb48b5603d8127f3e52f88d859e