building-lists.md
Summary:
The changes in this diff form part of the Strategic Review process initiated by Anna Murawska, which aims to improve the content, readability, and accuracy of the Litho documentation library.

building-lists.md: Removal of content while the page is marked as Under Construction.
start.mdx: minor change to opening sentence

Reviewed By: mihaelao

Differential Revision: D37749136

fbshipit-source-id: 6cf37337dc2266b8802f46d0e674475667f7ca9d