When attempting to mount with an empty rect during IM, rather than skip mount, proceed with mount with an empty rect
Summary: Currently, when incremental mount is enabled and we attempt to mount - we notice that the rect is empty and then do nothing. This means that any components / views that were previously mounted, just stay mounted. With this change, mounting when IM will proceed with the mount with a purely empty rect, thus ensuring that everything is removed as expected. Adjusted some tests to this new logic

Reviewed By: astreet

Differential Revision: D38151419

fbshipit-source-id: 02bff5a0510d947f414806536be3dc03f0cfb5c0