Support margin `auto` in KComponent Style
Summary:
`marginAuto` is supported on the Spec API but was omitted on the KComponent Style.
Looking for feedback on API direction. I had a few options in mind:

**A: `Style.margin(left = Dimen.auto, right = 10.px, top = Dimen.auto,..)`**

This introduces a special `auto` value that can be passed to margin instead of a Long value.

**Pro:**
- It's close to how Flex `auto` is passed directly to `margin`, and aligned with ReactNative and CK.

RN:

{F754873405}

CK:
{F754873372}

**Cons:**
- doesn't have parity with the Spec API
- we need a special `auto` value to pass for Dimen types

**B:  `Style.marginAuto(YogaEdge.LEFT).marginAuto(YogaEdge.TOP)`**

**Pro:**
- Same API as the Spec API
 {F754874233}

**Con:**
- I'm not a fan of cascading values for the same field, and I haven't seen this pattern anywhere else in the Kotlin APIs.
- It's not clear that this will be applied by Flexbox as a value on the `margin` and it's not a separate margin concept

**C:  `Style.marginAuto(YogaEdge.LEFT, YogaEdge.TOP)`**
Same as above, but instead of cascading for different edges, they can be passed as varargs

I like Option A the most and have started with that. This implementation of the `auto` value is a bit clunky, I bet there's a better way of doing this with Kotlin that I don't know, so I'd welcome any suggestions!

Reviewed By: astreet

Differential Revision: D38039959

fbshipit-source-id: a6c61c281aa7880aa60425ace555bd3b2e684db4