Release the checking for the children count of section
Summary:
This is a fix for the crash of `IndexOutOfBoundsException` in the `RecyclerBinder`, we would calculate a new changeset if the children count of section has changed.
Based on the result of [QE](https://fburl.com/deltoid3/dato4uzp) and the crash [log](https://fburl.com/logview/npinfmu9), seems like we can release it.

Reviewed By: astreet

Differential Revision: D38199648

fbshipit-source-id: 4297effab989f0acf257087d7645b2a8dbb85d66