Fix the bug of not propagating up the state of excludeFromIncrementalMount to its parent
Summary:
## Problem
Since we need to ensure all ancestors of the component that marked as `excludeFromIncrementalMount` to be mounted, but the previous implementation only did one level up, which may cause the entire view heirarchy get unmounted in some cases.

## Fix
Mark `needsUpdate` as `true` to call `ensureHostBounds` recursively, make sure all parents are marked as `excludeFromIncrementalMount` as well.

Reviewed By: adityasharat

Differential Revision: D38237503

fbshipit-source-id: b39e6dedb07486d29ae25415032e1ebe44dcc780