check nullability of nested diff node to check if we can use it
Summary:
We set diff node for nested tree in LSC and also maintain a boolean to see if it is being set. This started causing problems when we resolve OnCreateLayoutWithSizeSpec in measure phase always in next diff (Right now we resolve it in render phase when it is root node)

We were setting "null" diff node but still ended by setting the boolean to true also. I think we can get rid of boolean value and rely on nullability of nested diff node object

Reviewed By: astreet

Differential Revision: D38244646

fbshipit-source-id: 081ef0cf0ad70a1d99f2370f1752aa8850238d4a