Introduce MeasuredResultCache, allowing to cache measure results by component ID and LithoNode
Summary:
Currently, LayoutState is managing the layout result caching for both create and measure phase. This needs to change for splitting create and measure, so that layout results can be accessed either via component ID (as is the case today), or via LithoNode, which will be the approach during measure.

Introduced new class to managed this cache, and placed inside both RenderStateContext & LayoutStateContext. Follow-up diffs will remove this functionality from LayoutState.

Reviewed By: astreet

Differential Revision: D38206290

fbshipit-source-id: 5c676af7822e04848152cdb7fd0185d3f8639dba