Add API for resetting layout info
Summary:
## Problem

IG Reels is creating a new RecyclerView each time whenever users switch back from other tabs, which means we need to obtain the new LayoutManager of RecyclerView and pass it to the ReyclcBinder, otherwise, RecyclerBinder couldn't monitor the correct scrolling events while user is scrolling the list. Right now, ReyclerBinder only accepts LayoutManager with Builder and it cannot be reset later.

## Solution

We need to expose a new API, say `resetLayoutInfo`, to give us an oppertunity to reset the layout info.

Reviewed By: fabiocarballo

Differential Revision: D37970587

fbshipit-source-id: ec01a875daf1e36a078eb6c5bebbe987f9b12366