Enable Conditional Click/Touch Actions by Specifying If Enabled
Summary:
## This Diff

This diffs adds the ability to denote whether or not actions (click, long click, touch, intercept touch) actions should be added on the component for the KComponent API.  This, in essence, emulates the following API when working with event handlers:

```
.clickEventHandler(someBoolean ? MyComponent.onClick(c) : null)
```

Currently adding the above API does not exist for the KComponent API, since the fields take non-null lambdas.  This adds the ability to add a boolean flag for whether the action is disabled or enabled, which will supposed complex booleans that may not solely depend on lambda action nullability.

In a stacked diff, I'll remove the `onClick` without enabled overload.  In order to that, however, I need to clean up some of the existing code across the codebase to properly account for the boolean field.

----------------

Reviewed By: astreet

Differential Revision: D36619991

fbshipit-source-id: 49843760dbdd8ccc43a92bbd674aa30d491e63ee