Remove LithoLayoutResultHolder
Summary:
We already know from RenderStateContext whether the layout calculation was interrupted or not and we don't interrupt when we are in measure phase.
We can return early as we don't need LayoutResultHolder with interrupted result. This will also help to reduce redundancy of multiple methods like measureTree and layout (fixed in next diff)

Differential Revision: D38402409

fbshipit-source-id: e4eec20b9f67d7fa915b4586f1bb2b6426f0a8f4