Remove LayoutState from LayoutStateContext (1)
Summary: Multi-diff effort to remove LayoutState reference from LayoutStateContext to allow for further splitting. In this diff, we remove the need to fetch the LayoutState from the LayoutStateContext in InternalNodeUtils by passing the LayoutState thru the createRenderUnit methods. Since all the call-sites for createXXRenderUnit come from LayoutState to begin with, this change is fairly straight-forward.

Reviewed By: SidharthGuglani

Differential Revision: D38355153

fbshipit-source-id: 37086c56f1bb2b8e9fea6c9ac41cec0bff956e57