Remove LayoutState from LayoutStateContext (3)
Summary:
Introduced new interface - LayoutProcessInfo - used to get read-only info during the create / measure process rather than using the LayoutState itself.
Right now, LayoutState implements this interface, and it's still passed as the reference from LayoutStateContext - but in the future we'll be able to split this out and use a different object to read these values.

Currently only used for isCreateLayoutInProgress and getLayoutVersion.

Note:
isCreateLayoutInProgress is used to tell us if either create or measure is happening. This is needed to prevent too many state updates from happening during create / measure. In the future, we'll be able to replace this boolean by simply checking if there exists a RenderStateContext or LayoutStateContext that is not released. For now, we must still receive this indication from LayoutState.

getLayoutViesion is more tricky. Right now it seems to be only used during onMeasure / onBoundsDefined, and only from SizeSpecMountWrapperComponentSpec which AFAICT is only used in tests - so it's possible we might not need this at all. Unless there are objections or concerns, will leave this behaviour as is for now, and will re-examine the need for this in a later diff.

Reviewed By: SidharthGuglani

Differential Revision: D38355711

fbshipit-source-id: 41ce91b1e32c7706a8e4f1366bbfbe650397eac4