Remove LayoutState from LayoutStateContext (6)
Summary:
Removed requirement of LayoutState reference in Component.measure. Instead of creating a temp LayoutState, then extracting the LSC from it - we generate a temp LayoutStateContext directly which does not have a LayoutState reference at all.

Also marked the existing ctor for LSC as Deprecated, and created a new one that accepts an ID-generator and LayoutPhaseInfo instance instead.

The new TempLayoutStateContext uses this new ctor with trivial implementations for these interfaces.

Reviewed By: astreet

Differential Revision: D38386863

fbshipit-source-id: be02565cde6ca79cdad9e12af987442135d7a397