Remove RenderStateContext from LayoutStateContext (0)
Summary:
RenderStateContext's existence in LayoutStateContext was always intended to be a temporary step in an effort to separate Create & Measure. In the end state, RenderStateContext will be generated by the sequence that triggers the render phase (probably in ComponentTree). It will be passed thru the required places via method APIs where possible, or via Component-Context where necessary.

In this diff, we simply mark getRenderStateContext in LSC as deprecated to discourage further usage while the refactor progresses.

Reviewed By: astreet

Differential Revision: D38425975

fbshipit-source-id: a8060359bb4145bdddaa2ac2b2a727e8d899831f