Removed Layout-Version from LayoutProcessInfo & LayoutState
Summary:
Since the layout-version is static and never changes in a render / measure pass - we can hold it directly on LayoutStateContext rather than LayoutState - therefore it can also be removed from LayoutPhaseInfo, leaving on isCreateLayoutInProgress to be removed once RSC & LSC are separated.

Note: added static-final field to ComponentTree for INVALID_LAYOUT_ID to stay aligned throughout the code on what an invalid ID is. Minor refactors added in CT to use this global rather than hard-coded -1

Reviewed By: astreet

Differential Revision: D38425143

fbshipit-source-id: 83141078e7b9a7d33bc7e01fc744d980f300ef77