Expose the API of `setItemViewCacheSize` for RecyclerBinder
Summary: We had some issues with `ItemViewCache` before and disabled it in the [SectionsRecyclerView], but we need this feature to pre-bind offscreen items for full-screen size surfaces. So we override the value once `ItemPrefetch` is enabled and would remove these lines in the [SectionsRecyclerView] after we ensure everything is fine.

Reviewed By: fabiocarballo

Differential Revision: D38347202

fbshipit-source-id: 1901e9f27f14dffdc87b7882291d1d7bc24474c0