implementation of hasChildLithoViews in Mountable
Summary:
we expose boolean on RenderUnit that can be shared between litho and rendercore.
If your Mountable is rendering any render tree host children you need to override this as we are doing in VerticalScrollView and return true.
```
  override fun doesMountRenderTreeHosts(): Boolean = true
```
This will make sure that we notify children about the visible bounds changes

Reviewed By: colriot

Differential Revision: D38272888

fbshipit-source-id: e9f26f4685591ecd143f2d70289d0eb1770517d2