Remove willRender cache check for root component
Summary:
I think we will never hit the will render cache for root component. Since we are talking about the **root** component, that means will render must have been called on it outside layout calculation.

And in Layout calculation we will not have access to the same LSC used for will render in the first place.

Reviewed By: astreet

Differential Revision: D38625740

fbshipit-source-id: 9ce9e245f844e4cf1f494ec8164f6edf3626909a