Add Helpful Error When RecyclerView Not Mounted
Summary:
It took me a while to find what the issue was when I was using the API for the first time in one of my diffs.  Therefore, this adds a more helpful error in case other developers ever hit this same issue.

I'm not sure if there's a secret to know what Litho buck deps need to be added, so let me know if I've added more than necessary, potentially (it feels really hit or miss right now when I try to figure out what to add).
--------

Reviewed By: fabiocarballo

Differential Revision: D39068537

fbshipit-source-id: 48c02debc37d3fc8a7ea0c3a41747f6406698d8c