Remove LayoutStateContext from LithoLayoutResult & NestedTreeHolderResult
Summary:
We want to avoid holding a LSC inside a LithoLayoutResult, since in the future, the LithoLayoutResults may be created with different "temp" instances of LSC.

Now, instead of relying on the LSC that the result was created with, we use the provided LSC where needed.

To achieve this, we've removed the requirement when resolving nested trees that the LSC instance be equal to the one the result was created with. This should be fine, since if the layout pass is different, we won't have access to the cache in the 1st place anyway.

Reviewed By: astreet

Differential Revision: D39260449

fbshipit-source-id: ce401f102c7957498cb98d4e9c95ffe18a64e971