Convert all Layout, Component, LithoNode APIs to accept an RSC where applicable
Summary:
All render methods such as create, resolve, etc have been converted to accept a RenderStateContext rather than an LayoutStateContext.

To achieve this, RSC was given a pointer to LSC so that a scoped-component-context can be created with an RSC, and then set the LSC on itself. This introduces a cyclical reference, but will be removed once nested contexts are implemented.

All tests have been adjusted.

Reviewed By: astreet

Differential Revision: D39270593

fbshipit-source-id: 4b8c3c06b45c3f14140dcadba941d8a27c7bb504